requirements.txt
# AI Prompt Generator Platform
# Python 3.12+

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5

# AI Services
openai==1.12.0
google-generativeai==0.3.2
httpx==0.27.2

# Vector DB & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.5
langchain-openai==0.0.5
langchain-google-genai==0.0.6

# Google Sheets
gspread==5.12.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2

# Database
SQLAlchemy==2.0.25

# Utilities
python-dotenv==1.0.0
requests==2.31.0
validators==0.22.0
pydantic==2.5.3

# Dev Tools
pytest==7.4.3
black==23.12.1
flake8==7.0.0
